+++ bin/cli/CMakeLists.txt
@@ -4,11 +4,15 @@ cm_find_package(CM)
 include(CMDeploy)
 include(CMSetupVersion)
 
+if(CMAKE_SYSTEM_NAME STREQUAL "Android")
+    cm_project(devote WORKSPACE_NAME ${CMAKE_WORKSPACE_NAME} LANGUAGES ASM C CXX)
+else()
 cm_project(cli WORKSPACE_NAME ${CMAKE_WORKSPACE_NAME} LANGUAGES ASM C CXX)
+endif()
 
 if(NOT CMAKE_CROSSCOMPILING)
     cm_find_package(Boost COMPONENTS filesystem log log_setup program_options thread system)
-elseif(CMAKE_CROSSCOMPILING)
+elseif(CMAKE_CROSSCOMPILING AND CMAKE_SYSTEM_NAME STREQUAL "Emscripten")
     if(NOT TARGET boost)
         include(ExternalProject)
         set(Boost_LIBRARIES boost_random)
@@ -23,6 +27,11 @@ elseif(CMAKE_CROSSCOMPILING)
     else()
         set(Boost_LIBRARIES boost_random)
     endif()
+elseif(CMAKE_CROSSCOMPILING AND CMAKE_SYSTEM_NAME STREQUAL "Android")
+    set(Boost_INCLUDE_DIRS "${CMAKE_SOURCE_DIR}/../boost/${ANDROID_ABI}/include/boost-1_78")
+    set(Boost_LIBRARIES ${CMAKE_SOURCE_DIR}/../boost/${ANDROID_ABI}/lib/libboost_random.a)
+    find_library(log-lib log)
+    list(APPEND PLATFORM_SPECIFIC_LIBRARIES ${log-lib})
 endif()
 
 cm_setup_version(VERSION 0.1.0)
@@ -31,13 +40,29 @@ cm_setup_version(VERSION 0.1.0)
 # e.g., for creating proper Xcode projects
 set(${CURRENT_PROJECT_NAME}_HEADERS)
 
-# list cpp files excluding platform-dependent files
+if(CMAKE_SYSTEM_NAME STREQUAL "Android")
+    list(APPEND ${CURRENT_PROJECT_NAME}_SOURCES
+            src/android.cpp)
+
+    add_library(${CURRENT_PROJECT_NAME}
+            SHARED
+            ${${CURRENT_PROJECT_NAME}_HEADERS}
+            ${${CURRENT_PROJECT_NAME}_SOURCES})
+elseif(CMAKE_SYSTEM_NAME STREQUAL "Emscripten")
+    list(APPEND ${CURRENT_PROJECT_NAME}_SOURCES
+            src/wasm.cpp)
+
+    add_executable(${CURRENT_PROJECT_NAME}
+            ${${CURRENT_PROJECT_NAME}_HEADERS}
+            ${${CURRENT_PROJECT_NAME}_SOURCES})
+else()
 list(APPEND ${CURRENT_PROJECT_NAME}_SOURCES
      src/main.cpp)
 
 add_executable(${CURRENT_PROJECT_NAME}
                ${${CURRENT_PROJECT_NAME}_HEADERS}
                ${${CURRENT_PROJECT_NAME}_SOURCES})
+endif()
 
 set_target_properties(${CURRENT_PROJECT_NAME} PROPERTIES
                       LINKER_LANGUAGE CXX
@@ -59,7 +84,9 @@ target_link_libraries(${CURRENT_PROJECT_NAME}
                       marshalling::crypto3_algebra
                       marshalling::crypto3_zk
 
-                      ${Boost_LIBRARIES})
+                      ${Boost_LIBRARIES}
+
+                      ${PLATFORM_SPECIFIC_LIBRARIES})
 
 target_include_directories(${CURRENT_PROJECT_NAME} PUBLIC
                            $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/include>
